requirements.txt
pywebview>=4.4.1
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
